requirements.txt
flask
pyyaml
openai-whisper
yt-dlp
nltk
ffmpeg-python
pillow
TTS
moviepy
python-dotenv
numpy
torch